requirements.txt
azure-functions==1.23.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
iniconfig==2.1.0
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
